requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
firebase-admin
python-multipart
